Add random pick function and related documentation
This commit adds the random pick function in 'main.go' that picks an index based on the weights provided. It uses a sum prefix array to calculate the cumulative weights and then randomly picks an index. A corresponding README file is also added to provide a detailed explanation of how this functionality works and some example uses.