feat: add next permutation functionality
This commit introduces a new function `nextPermutation()` in the `main.go` file for finding the next lexicographically greater permutation of an array of integers. It also provides a new `README.md` which contains the problem description, examples and constraints related to this function. This functionality has a constant space complexity as it operates in-place.